delete function for only corresponding user

+++ ExepenseAppControllers/addExpense.js
@@ -2,9 +2,7 @@ const Expense=require('../ExpenseAppModels/expense');
 const jwt=require('jsonwebtoken')
 exports.addExpense=async(req,res,next)=>{
     try{
-        const token=req.header('Authorization')
-        const user=jwt.verify(token,'poojasecretkey')
-              
+        
         if(!req.body.amount||!req.body.description||!req.body.category){
             console.log('please fill all the details')
             res.status(500).json({err:'not filled'})        }
@@ -12,7 +10,7 @@ exports.addExpense=async(req,res,next)=>{
         const description=req.body.description;
         const category=req.body.category;
         
-        const expense=await Expense.create({amount:amount,description:description,category:category,userId:user.userId})
+        const expense=await Expense.create({amount:amount,description:description,category:category,userId:req.user.id})
        
         res.status(201).json({expense})
     }
@@ -23,7 +21,7 @@ exports.addExpense=async(req,res,next)=>{
 exports.getExpense=async(req,res,next)=>{
     try{
     
-        const expenses=await Expense.findAll({where:{userid:req.user.id}})
+        const expenses=await Expense.findAll({where:{userId:req.user.id}})
         return res.status(200).json({Expenses:expenses})
     }
     catch(err){
@@ -36,8 +34,17 @@ exports.deleteExpense=async(req,res,next)=>{try{
         console.log('expense not found')
         res.status(500).json({err:'not found'})      
     }
+    
     const expenseId = req.params.expenseId;
-    await Expense.destroy({where:{id:expenseId}})
+    await Expense.destroy({where:{id:expenseId,userId:req.user.id}})
+    .then((response)=>{
+        if(response===0){
+            res.status(400).json({message:'expense can not be deleted as it belongs to other'})
+        }
+        else if(response===1){
+            res.status(200).json({message:'deleted successfully'})
+        }
+    })
     
   }
